Feat: 환전 API 구현
원화 -> 외화, 외화 -> 원화 환전 API 구현 및
기존 환전 관련 금액 자료형 int -> Long 수정

Fix: #feat/54_payment

+++ triplet/src/main/java/com/ssafy/triplet/exchange/dto/response/ExchangeRateCalculatorResponse.java
@@ -11,9 +11,9 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ExchangeRateCalculatorResponse {
 
-    private int sourceAmount;
+    private Long sourceAmount;
 
-    private int targetAmount;
+    private Long targetAmount;
